requirements.txt
PyJWT==2.6.0
Flask==2.2.3
Werkzeug==2.2.3
cryptography==39.0.1
requests==2.28.2
python-dateutil==2.8.2
tabulate==0.9.0

# Optional dependencies
colorama==0.4.6  # Optional: For colored terminal output